# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/field-calculations": {
      "post": {
        "operationId": "FieldCalculationController_create",
        "summary": "Save a field calculation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFieldCalculationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Calculation saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldCalculationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          }
        },
        "tags": [
          "field-calculations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "FieldCalculationController_findAll",
        "summary": "Get all field calculations",
        "parameters": [
          {
            "name": "workOrderId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of calculations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldCalculationResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "field-calculations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/field-calculations/technician/{technicianId}": {
      "get": {
        "operationId": "FieldCalculationController_findByTechnician",
        "summary": "Get calculations by technician",
        "parameters": [
          {
            "name": "technicianId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of calculations by technician",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldCalculationResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "field-calculations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/field-calculations/{id}": {
      "get": {
        "operationId": "FieldCalculationController_findOne",
        "summary": "Get a specific field calculation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calculation found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldCalculationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Calculation not found"
          }
        },
        "tags": [
          "field-calculations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "FieldCalculationController_delete",
        "summary": "Delete a field calculation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calculation deleted successfully"
          },
          "404": {
            "description": "Calculation not found"
          }
        },
        "tags": [
          "field-calculations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/field-calculations/{calculationId}/attach": {
      "put": {
        "operationId": "FieldCalculationController_attachToWorkOrder",
        "summary": "Attach calculation to work order",
        "parameters": [
          {
            "name": "calculationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calculation attached successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldCalculationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Calculation not found"
          }
        },
        "tags": [
          "field-calculations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/feedback": {
      "post": {
        "operationId": "FeedbackController_create",
        "summary": "Submit feedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "feedback"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "FeedbackController_findAll",
        "summary": "List all feedback (Admin only)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "feedback"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/feedback/{id}": {
      "get": {
        "operationId": "FeedbackController_findOne",
        "summary": "Get feedback by ID (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "feedback"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "FeedbackController_update",
        "summary": "Update feedback status (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "feedback"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "FeedbackController_remove",
        "summary": "Delete feedback (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "feedback"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/feedback/stats/summary": {
      "get": {
        "operationId": "FeedbackController_getStats",
        "summary": "Get feedback statistics (Admin only)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "feedback"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "HVAC API",
    "description": "Generated API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateFieldCalculationDto": {
        "type": "object",
        "properties": {
          "calculatorType": {
            "type": "string",
            "description": "Calculator type/name",
            "example": "Ohm's Law"
          },
          "category": {
            "type": "string",
            "description": "Calculator category",
            "enum": [
              "electrical",
              "refrigeration",
              "airflow",
              "gas",
              "hydronic",
              "utility"
            ],
            "example": "electrical"
          },
          "inputs": {
            "type": "object",
            "description": "Calculator input data as JSON",
            "example": {
              "voltage": 120,
              "current": 10,
              "resistance": 12
            }
          },
          "results": {
            "type": "object",
            "description": "Calculator result data as JSON",
            "example": {
              "V": 120,
              "I": 10,
              "R": 12,
              "P": 1200
            }
          },
          "workOrderId": {
            "type": "string",
            "description": "Work order ID to attach calculation to",
            "example": "wo-uuid-123"
          },
          "notes": {
            "type": "string",
            "description": "Optional notes about the calculation",
            "example": "Measured at main panel"
          }
        },
        "required": [
          "calculatorType",
          "category",
          "inputs",
          "results"
        ]
      },
      "FieldCalculationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "technicianId": {
            "type": "string"
          },
          "calculatorType": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "inputs": {
            "type": "object"
          },
          "results": {
            "type": "object"
          },
          "workOrderId": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "isAttached": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "technician": {
            "type": "object"
          },
          "workOrder": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "tenantId",
          "technicianId",
          "calculatorType",
          "category",
          "inputs",
          "results",
          "isAttached",
          "createdAt",
          "updatedAt",
          "technician"
        ]
      },
...
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "bug",
              "feature_request",
              "complaint",
              "praise",
              "other"
            ]
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "message": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "pageUrl": {
            "type": "string"
          }
        },
        "required": [
          "category",
          "message"
        ]
      },
      "UpdateFeedbackDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "new",
              "reviewed",
              "in_progress",
              "resolved",
              "dismissed"
            ]
          },
          "adminNotes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}